Obsetvers for Focusing and Slaying Cards + Behavior in Environment
Added another observer - SlayListener:
 * used to check if enemy or environment is slain
 * when enemy slain  //should be added later, not in the code rn - generate another environment

When a focus change is registers the notifyIsActiveListeners checks the type of the card that called the notification and responds correspondingly.

When a slay() registers the notifySlayListeners checks the type of the card that called the notification and responds correspondingly.

IN ENVIROMENT:
When dagger has been focused slay the last focused card. Different cards react differently on being slain.
 * When the player is slain the application ends